fix(attachments): include older xpaths in data api DEV-596 (#5862)
### 📣 Summary
Ensure we show the correct xpath for attachments submitted under older
versions of a survey..


### 📖 Description
There is a bug in KPI wherein if an image question was moved into a
group and the survey had responses from before and after the move, a red
exclamation mark would show up in cells containing attachments from
before the move. Part of the problem was that, for these attachments,
the API was returning empty xpaths. The remaining part of the problem is
in the front-end, but this PR fixes the API response.

### 💭 Notes
Updates the deployment backends to fetch all attachment xpaths using a
new `get_all_attachment_xpaths` that goes through all versions of an
asset and gets the xpaths of all attachment questions. Passes the result
to `_rewrite_json_attachment_urls` so we don't have to call the method
again for every submission. Removes the old `get_attachment_xpaths`
method and replaces all usages with passing the results of the new
method.