fix(dataTable): prevent 500 error when opening data table view DEV-695 (#5904)
### 📣 Summary
Prevent a 500 error when opening data table view when the project
contains versions created before the NLP feature.

### 📖 Description
This PR resolves a 500 Internal Server Error that could occur when
opening the data table view for projects with old versions.
The issue was caused by a TypeError when the system attempted to access
a `$xpath` property that doesn’t exist on versions created/deployed
before the introduction of the NLP feature.