docs(api): adding API endpoints for asset-data in KPI documentation DEV-527 (#5867)
### 📣 Summary
Added API documentation for `/api/v2/docs/asset_data` endpoint.


### 📖 Description
This PR introduces the endpoint documentation that will be used for the
auto-generating API documentation of the asset_attachments endpoint.
With these changes, users will be able to understand, test and use the
`asset_data` endpoint more easily.


### 💭 Notes
Due to current limitations in drf-spectacular support for OAS 3.1, it's
not possible to include structured request or response bodies for
certain endpoints. This applies to all functions flagged as "vague."
As a workaround, example payloads have been included directly in the
description fields of the affected endpoints. Full support is pending a
future update of the tooling.

### 👀 Preview steps
To see the changes, visit `http://kf.kobo.local/api/v2/docs/`. The page
of the Kobo API will now be visible and generated by swagger. To see the
changes added to `asset_data` endpoint, scroll down to the category of
the same name or make a research in the search bar. Documentation for
the endpoint will be provided with a HTTP code, a body and/or a response
when necessary.