docs(api): adding API endpoints for users in KPI documentation DEV-549 (#5927)
### 📣 Summary
Added API documentation for `/api/v2/docs/users` endpoint.

### 📖 Description
This PR introduces the endpoint documentation that will be used for the
auto-generating API documentation of the users endpoint. With these
changes, users will be able to understand, test and use the `users`
endpoint more easily.

### 👀 Preview steps
To see the changes, visit `http://kf.kobo.local/api/v2/docs/`. The page
of the Kobo API will now be visible and generated by swagger. To see the
changes added to `users` endpoint, scroll down to the category of the
same name or make a research in the search bar. Documentation for the
endpoint will be provided with a HTTP code, a body and/or a response
when necessary.