docs(api): adding API endpoints for social in KPI documentation DEV-832 (#6047)
### 📣 Summary
Added API documentation for `/me/social/` endpoint.

### 📖 Description
This PR introduces the endpoint documentation that will be used for the
auto-generating API documentation of the `social` endpoint. With these
changes, users will be able to understand, test and use the `social`
endpoint more easily.

### 👀 Preview steps
To see the changes, visit `http://kf.kobo.local/api/v2/docs/`. The page
of the Kobo API will now be visible and generated by swagger. To see the
changes added to `social` endpoint, scroll down to the `Me` category or
make a research in the search bar. Documentation for the endpoint will
be provided with a HTTP code, a body and/or a response when necessary.