docs(api): refactor of drf_spectacular schema DEV-727 (#5995)
### 📣 Summary
Refactor of drf_spectacular code for better structure and cleanliness.

### 📖 Description
Make sure the files for documenting the KPI `api/v2` follows the best
practices establish during development. This include URLS, serializers,
array building and schema examples.


### 👀 Preview steps
1. Go to `http://kf.kobo.local/api/v2/docs/`
2. Look at all endpoints url and approve the followings:
- Urls are correctly represented
- Responses and payloads still make senses
3. For each endpoints, approve the followings:
- The serializers extensions (when an endpoint uses its own serializer
and not an inline_class_serializer) have classes extending the fields
that require them or @extend_schema_field on a get_[...] method for
required fields.
- In the `schema_extension/api/v2/[endpoint]/extensions.py` file for
each endpoint, make sure that, if we are building a `build_array_type()`
that there is a schema inside the util. It must not be empty and must at
the very least have the following:
`schema=build_basic_type(OpenApiType.STR)`. This is to stop schema
errors from raising.
- **IF** the endpoint URL you are viewing has many examples (most
commonly oneOf or anyOf in CREATE or PATCH endpoints) make sure we are
not extending the schema of its viewset and are instead giving examples
directly in the schema_extension.
4. Open a terminal where you check the logs of the kpi docker. Reload
the doc page and make sure there are no errors or warnings popping up.
5. Download the schema.yaml of the documentation and slide it in
`https://oas-validation.com/`. All validations must pass.

---------

Co-authored-by: Olivier Leger <olivierleger@gmail.com>