docs(api): move OpenRosa endpoints to separate docs DEV-774 (#6082)
### 📣 Summary
Separate OpenRosa and `v2` API endpoints to their own documentation.

### 📖 Description
To ensure a better a documenting of our apis, we need to separate them
into their respective schemas.

### 👀 Preview steps
1. Visit `http://kf.kobo.local/api/v2/docs/`, this should lead to the
`v2` endpoints documentation. Make sure there are no OpenRosa or `v1`
endpoints present. Make sure you can download the schema and that it is
the right one.
1. Visit `http://kf.kobo.local/api/openrosa/docs/`, this should lead to
the OpenRosa endpoints documentation. Make sure there are no `v2`or `v1`
endpoints present. Make sure you can download the schema and that it is
the right one.

---------

Co-authored-by: Olivier Leger <olivierleger@gmail.com>