feat(auditLogs): optimize search filter for audit log endpoints (#6129)
### 📣 Summary
Performance enhancement for audit log, access log, and project history
log endpoints.


### 💭 Notes
The SearchFilter automatically runs `distinct` on all queries in order
to handle the possibility that searching on an n-to-many field has
returned multiple results for the same object. This distinct is very
slow, so add a property to the view that tells the SearchFilter to skip
it. This property should only be set on models that do not have any kind
of n-to-many fields.

### 👀 Preview steps
This is best previewed by running the server with `--print-sql` or
turning on the debug toolbar

1. ℹ️ have an account
2. Navigate to `api/v2/audit-logs/?q=user__username:<username>
3. 🟢 [on PR] Notice the audit log query does not use `distinct` for the
search or for the count