feat(projectOwnership): add management command for retrying attachment transfers DEV-842 (#6121)
### 📣 Summary
Add a process for retrying the transfer of attachments after a previous
failure.


### 📖 Description
Add a management command that can be run to retry the transfer of
attachments between old and new owners of a project after an initial
failure.


### 💭 Notes
The management command takes a Transfer uid since `move_attachments`
expects a transfer as the parameter. This also ensures that we are only
trying to copy attachments to the desired recipient. The PR also adds a
generic try/catch around the processing of each attachment as well as
some more logging. This is to make it more easy to determine which
attachments, if any, are failing and why. Once we have a better idea of
the problems occurring we can catch more specific exceptions and remove
the logging if it is getting excessive.

### 👀 Preview steps
This is most easily tested if you are using an S3 backend.

1. ℹ️ have an MMO with at least two members
2. ℹ️ have at least one asset with submissions with attachments in the
MMO
3. Update AWS_SECRET_ACCESS_KEY to be something incorrect (this will
ensure the attachment transfers fail the first time around)
4. In Django admin, update the organization owner to a different user
5. In Django shell, verify that the relevant TransferStatus
(status_type='attachments') failed
6. Update the AWS_SECRET_ACCESS_KEY to the correct one and restart the
kpi_worker
7. In a kpi container, run `./manage.py retry_attachment_transfers
<transfer_uid>`
8. 🟢 The attachments should now be transferred to the new location, or
have logged the relevant errors