refactor(UniversalTable): handle future generated API helpers DEV-895 (#6134)
### 💭 Notes

Restructured UniversalTable to handle generated API helpers in the
future. It should be possible to swap queries without more changes to
the UniversalTable. Notably (see commit by commit):
- Extracted the pagination state and query hook to parent component for
simplicity. The future generated API hooks and fetchers would replace
all of the `*.query` file contents. With previous `queryHook` approach
it would require an extra boilerplate to re-wrap each hook to conform to
the `queryHook` signature and complex typing gymnastics (I tried and
failed), therefore I settled for an equal amount of boilerplate of
raising pagination state along much simpler UniversalTable types and
removing the mental complexity of passing a hook as a parameter.
- unrelated but along the way, renamed UniversalTable files according to
the frontend coding style guidelines. Note that nowhere UniversalTable
was used directly, so I renamed the PaginatedQueryUniversalTable to
UniversalTable and UniversalTable to UniversalTableCore as an internal
component.


### 👀 Preview steps
1. ℹ️ have an account and a project
2. 🟢 [on PR] notice that organization member table still works
3. 🟢 [on PR] notice that form activity table still works as before
4. 🟢 [on PR] notice that access logs table still works