feat(components): add tooltips to icon only buttons with storybook option DEV-359 (#6123)
### 📣 Summary
Adds ability for tooltips to be used with icon only buttons. Useful for
some UI designs

### 💭 Notes
After https://github.com/kobotoolbox/kpi/pull/5795 we realized there's
not icon only button with a tooltip so we used a deprecated `Button`
component instead. This PR replaces it with the icon only button with
the new tooltip (`ActionIcon`).
 
### 👀 Preview steps
1. ℹ️ have an account and a project
2. make multiple versions of the project
3. on the form summary page, click "show all versions"
4. 🟢 [on PR] notice that the clone icons are the same
5. Go the `ActionIcon` story in storybook
6. 🟢 [on PR] notice that there is a prop for tooltips and it works demos
as expected