ci(api): check generated openapi DEV-859 (#6119)
### 💭 Notes

For development purposes let's commit the generated schema, so that we
can build on top of that frontend typed API helpers, see #6083.

To ensure that the committed schema is always up-to-date, let's add to
CI a job that checks if it's generated.

P.S. along the way simplified the script with a helper for gosu.

### 👀 Preview steps

1. see the new "openapi" job
2. 🟢 [on PR showcase commit
[f4d9501](https://github.com/kobotoolbox/kpi/pull/6119/commits/f4d9501a729f2224f777a37e0914ca33e0d47b2b)
before updating schema] notice that CI fails
3. 🟢 [on PR latest commit
[936c461](https://github.com/kobotoolbox/kpi/pull/6119/commits/936c461989f88e5f277baf6cf77d2bbdac57c293)
after updating schema] notice that CI succeeds
4. 🟢 run locally `./run.py -cf run --rm kpi
./scripts/generate_openapi_schemas.sh` from `kobo-install` and it
succeeds to regenerate schema