docs(api): adding openrosa endpoints for submission in KPI documentation DEV-771 (#6109)
### 📣 Summary
Added API documentation for `/api/openrosa/docs/submission` endpoint.

### 📖 Description
This PR introduces the endpoint documentation that will be used for the
auto-generating API documentation of the `submission` endpoint. With
these changes, users will be able to understand, test and use the
`submission` endpoint more easily.

### 👀 Preview steps
To see the changes, visit `http://kf.kobo.local/api/openrosa/docs/`. The
page of the Kobo API will now be visible and generated by swagger. To
see the changes added to `submission` endpoint, scroll down to the
category of the same name or make a research in the search bar.
Documentation for the endpoint will be provided with a HTTP code, a body
and/or a response when necessary.