docs(api): adding openrosa endpoints for formList in KPI documentation DEV-877 (#6095)
### 📣 Summary
Added API documentation for `/api/openrosa/docs/formList` endpoint.

### 📖 Description
This PR introduces the endpoint documentation that will be used for the
auto-generating API documentation of the `formList` endpoint. With these
changes, users will be able to understand, test and use the `formList`
endpoint more easily.

### 👀 Preview steps
To see the changes, visit `http://kf.kobo.local/api/openrosa/docs/`. The
page of the Kobo API will now be visible and generated by swagger. To
see the changes added to `formList` endpoint, scroll down to the
category of the same name or make a research in the search bar.
Documentation for the endpoint will be provided with a HTTP code, a body
and/or a response when necessary.

---------

Co-authored-by: Olivier Leger <olivierleger@gmail.com>