feat(dataCollectors): set up initial models and admin DEV-786 (#6142)
### 📣 Summary
Add new data collector models to Django admin.

### 📖 Description
Adds DataCollector and DataCollectorGroup models to Django admin. A
DataCollectorGroup may have several DataCollectors and be associated
with several Assets. Every DataCollector has a token, which may be
rotated using Django admin.

### 💭 Notes
The admin for DataCollectors and DataCollectorGroups is extremely
rudimentary because it turns out making a nice interface for one-to-many
relationships is hard. For this PR, the user can add assets to a
DataCollectorGroup by selecting from a giant checklist of all available
assets. Hopefully in the future, this will be enhanced with an
autocomplete that allows them to search for assets before adding them.
Also in the future, users will be able to add DataCollectors on the
DataCollectorGroup page rather than having to go add the group to the
DataCollector objects. These efforts are part of DEV-920 in Linear.

For now, making sure the owner of the DataCollectorGroup has
manage_asset permissions is only validated at the admin level. Later on
we will probably want to do so with the serializer or even on the model
itself but that raises questions about what to do when the owner of a DC
group changes or is removed which we do not need to worry about for the
pre-MVP. For now, we just filter the list of assets shown down to ones
the DC group owner has access to if the owner is already set, and raise
validation errors if the owner changes to one without the proper
permissions. We also filter the list to assets that don't already have a
DC group set since each asset can only have one.

The new classes were added to kobo_auth since they are a type of user
(sort of) and we mostly care about them for authentication purposes.


### 👀 Preview steps
Make sure to run migrations and undo them afterwards

1. ℹ️ have an admin account
2. In Django admin, create a new DataCollectorGroup
3. 🟢 Your admin user should be set as the default owner
4. Hit 'Save and continue editing'
5. 🟢 You should now be able to add assets to the group
6. 🟢 In a Django shell, validate that your DataCollectorGroup exists and
that all associated assets have `data_collector_group_id=<dcg_id>`
7. Add DataCollectors and set the group to your new DataCollectorGroup
8. 🟢 Notice all newly created DataCollectors have been assigned a token
9. 🟢 In a django shell, validate that `my_group.data_collectors.all()`
contains all the DataCollectors you added
10. In the DataCollectors list view, select several and choose the
action 'Rotate token'
11. 🟢 Notice all tokens for the selected DataCollectors have been
updated