feat(reversion): task to remove version objects DEV-946 (#6162)
### 💭 Notes
Background celery job to remove Version objects as we deprecate use of
the library. The version table is likely to be one of the largest in the
database so removing it all in one drop table command could be
prohibitively expensive. There are `sleep`s built in to avoid
overwhelming the table and the job will time out after 30m to avoid
eating the queue.


### 👀 Preview steps

1. ℹ️ Ensure you have > 3 Version objects (you can do so by editing
instances)
2. Set VERSION_DELETION_BATCH_SIZE to 2
3. Set the log level to debug
4. 🟢 [on PR] as the job runs, notice that versions are deleted 2 at a
time