feat(dataCollectors): utilities for creating enketo links DEV-798 (#6196)
### 💭 Notes
Developer-only changes. Creates the utils we will use to create the
correct enketo links for data collectors, separated out into its own PR
for easier reviewing. The actual processes that call these utils will be
updated in separate PRs.
We are replacing the default 8-char enketo_ids with longer hashes for
more security around data-collector endpoints since they do not require
login. The new ids are only 31 characters long because 32 characters
causes enketo to 404.

There is a quirk mentioned in
https://chat.kobotoolbox.org/#narrow/stream/4-Kobo-Dev/topic/DEV-798.20creating.20enketo.20links/near/684244
that is not handled in this PR because it didn't seem to affect the
functionality. It should probably be handled so we don't end up with
cruft in redis but I think it can be done separately.

### 👀 Preview steps

1. Open a django shell and a redis-cli
2. In the django shell, run `set_data_collector_enketo_links('token_a',
['uid_0', 'uid_1'])`
3. In redis, run `hget "or:<kobo_url>/key/token_a" "uid_0"` and make
note of the 31-char hash you get back
4. In a browser, navigate to <enketo_url>/x/<hash>
5. 🟢 You should get a response `Request to
http://kf.kobo.local:8080/key/token_a/formList?formID=uid_0 failed.`
This means it generated the correct request, but we haven't set up
authentication yet so we expect it to fail
6. Repeat steps 3-5 with "uid_1" 
7. In the django shell, run
`rename_data_collector_enketo_links('token_a', 'token_b')`
8. Navigate to the same urls you used in enketo earlier (using the 31
char hashes)
9. 🟢 You should get a response `Request to
http://kf.kobo.local:8080/key/token_b/formList?formID=uid_0 failed.`
(note the changed token)
10. In the django shell, run
`remove_data_collector_enketo_links('token_b', ['uid_1'])`
11. Navigate to the enketo url you used for `uid_1`
12. 🟢 You should get a response `Survey with this ID not found` (because
the entry in redis for this token/asset has been removed)
13. In the django shell, run
`remove_data_collector_enketo_links('token_b')`
14. Navigate to the enketo url you used for `uid_0`
15. 🟢 You should get a response `Survey with this ID not found` (because
the entry in redis for the remaining token/assets have been removed)