ci(releases): generate changelog DEV-301 (#6150)
### 💭 Notes

Changelog is automatically generated and force-pushed to a corresponding
`changelog/*` branch to provide a consistent and easily discoverable
link. These changelog branches can be manually deleted at any time, e.g.
for housekeeping.


### 👀 Preview steps

1. see example output at
https://github.com/kobotoolbox/kpi/blob/changelog/2.025.34/CHANGELOG.md
as-if run on the `2.025.34` branch just before the release.
2. try `npm run changelog` to generate a changelog locally
3. eyeball
[release-2-stabilize.yml](https://github.com/kobotoolbox/kpi/pull/6150/files#diff-efd13573359fc36420aa5cc4c25df08c6fc0d5ffa93c4e0518ff99f87d8f71f8),
because can't really test without merging into a release branch.