fix(mapView): fix disaggregate button not working DEV-991 (#6218)
### 📣 Summary
This PR fixes an issue where the disaggregate button in the map data
view was not working.

### 📖 Description
- The issue was happening due to a small mistake in importing, which was
importing the exported class of the component without the `withRouter`
HOC instead of the `default` export from the file, causing an error when
trying to access the router component in the disaggregate function.
- I fixed the import and removed the export that should never be used
for safety.

### 👀 Preview steps
1. ℹ️ have an account
2. ℹ️ have a project with a "pick one" and a "point" questions
3. ℹ️ have some submissions with different values on both questions
4. Go to project Data > Map section
5. Click the bottom left `Disaggregate by survey responses` button
6. 🔴 [on main] throws an error and do not work
7. 🟢 [on PR] function works and allows for filtering points