feat(dataCollectors): create authentication class for data collectors DEV-787 (#6203)
### 💭 Notes
Developer-only changes. Sets up the authentication class we will use to
authenticate DataCollectors, who have authentication tokens but are not
actual Django users.
Since we need some kind of user to verify permissions, we create a new
"fake" user class, DataCollectorUser, that inherits from AnonymousUser.
This type of user has a list of assets it is allowed to view/submit to.
When we authenticate, we grab the token from the url and use it to look
up a DataCollector object. If there is one, we create a new
DataCollectorUser and give it a list of assets corresponding to the
assets belonging to the DataCollector's group. This DataCollecotrUser's
only permissions are PERM_ADD_SUBMISSION on any asset in its whitelist.

### 👀 Preview steps
None, dev-only.