Clear ./node_modules/.cache/storybook in CI
Run `rm -rf --verbose ./node_modules/.cache/storybook`, both before and
after the Storybook test suite runs.

In an ideal world, we would prevent Storybook entirely from using or
writing to its own cache in CI environment. But Storybook's caching
behaviors (and the mechanism for controlling them) could change between
Storybook updates.

So:

- Delete Before: Prevent uncaught cache pollution from affecting this
  run.
- Delete After: Prevent this run from polluting the node_modules cache.
- Use `--verbose` to clarify the extent of the Storybook cache in both
  cases.

If you're working on a Storybook PR, it might be a good idea to check
these logs in CI to see what the Storybook cache is attempting to write.