fix(sso): update primary email when enabling SSO DEV-1005 (#6250)
### 📣 Summary
Update a user's email if they enable SSO and do not allow them to change
it as long as SSO is enabled.

### 📖 Description
When a user connects their account to an SSO provider with an email
address, we should update their email address to match the email used by
the SSO. Once this is set, the user should not be allowed to change it.


### 💭 Notes
This PR makes the assumption that if a user has a SocialAccount and an
EmailAddress, the email address probably came from the SSO and so should
not be changed. There is an edge case where this wouldn't be true:
1. ACCOUNT_EMAIL_REQUIRED = False
2. The SSO provider does not use email
3. The user separately created and verified their email

However, this should be extremely rare, and it's very hard to determine
if an EmailAddress came from an SSO, so we're choosing to keep the
assumption in.

Besides preventing users from updating their emails if they have SSO
enabled, this PR also takes care of setting the correct email address
for a user if they enable SSO on a pre-existing account. It's somewhat
surprising allauth doesn't do this automatically, though it seems like
something they considered:
https://github.com/pennersr/django-allauth/blob/52e90f442ee4f5c2f15f609d130f4b61397b6f3c/allauth/socialaccount/models.py#L313.
In case we get multiple email addresses from the SSO provider, which is
possible, we use allauth's `cleanup_email_addresses` to select the
correct primary one and set all the other ones to not-primary. Then we
manually send the email_confirmed signal to indicate that we have
confirmed this new primary email, since by the time the
`social_account_added` signal is called, we know the email has been
confirmed. The existing handler for the signal takes care of cleaning up
the rest of the non-primary emails.

We do this instead of removing all the non-primary addresses just in
case we ever change the logic around what to do when we have a new
verified email and decide we want to allow >1 email per user.


### 👀 Preview steps
Enable SSO locally.

1. ℹ️ have an account and a project. Use a different email for the
account than you will use for SSO.
5. In Account Settings > Security, enable Single Sign-On
6. Connect your account with Kobo Google Apps
7. Return to Account Settings > Security
8. 🔴 [on release] The email address has not changed
9. 🟢 [on PR] The email address has changed to the SSO email
10. Try to change the email
11. 🔴 [on release] You are able to update the email
12. 🟢 [on PR] updating the email results in a 400 (this will not yet be
reflected in the UI, but if you reload the page you should see that the
email hasn't changed)