ci(releases): deploy to beta only newest release DEV-1060 (#6312)
### 💭 Notes

If a commit was pushed to the previous release branch that merges
forward into the newest release, that would trigger two deploys to beta.
Deploying previous release is unnecessary, as the newest release will
deploy over it shortly after. Furthermore with unlucky timing it
introduces risks for QA team to be confused what they are actually
testing when.