feat(dataCollectors): allow DCs access to read-only OpenRosa endpoints DEV-788 (#6258)
### 📣 Summary
Allow data collectors to view surveys in Enketo.


### 📖 Description
Updates the OpenRosa endpoints to allow urls of the form
`/key/<token>/endpoint` and provide the correct response if the data
collector with the relevant token is part of a group that has access to
the xform.


### 💭 Notes
Updates the read-only OpenRosa endpoints (list, manifest, media, and
download) to accept requests with the token in the URL, using the new
DataCollectorTokenAuthentication class and checking to make sure the
DataCollectorUser has access to the xform via their group.


### 👀 Preview steps

1. ℹ️ have an account and a project with at least one image
2. In django admin, create a data collector group with at least one data
collector
3. Assign the project to the data collector
4. In redis, get the enketo id with 'get
"or:kf.kobo.local:8080/key/<dc_token>,<asset_uid>"'
5. Go to http://ee.kobo.local/x/<enketo_id>
6. 🟢 [on PR] The survey should load. Note the image may not properly
load if using kobo-compose but that's a pre-existing problem
7. Check the kpi logs
8. 🟢 [on PR] There should be requests to
`/key/token/formList?formID=<xformid>`,
`/key/token/xformsManifest/<xform_pk>`, and
`/key/token/forms/<xform_pk>/form.xml`